store logic moved to a separate folder and refactored a bit
