+++ src/Root.jsx
@@ -3,7 +3,7 @@ import { createStore } from 'redux';
 import { Provider } from 'react-redux';
 
 import App from './components/App';
-import rootReducer from './reducers';
+import rootReducer from './store/reducers';
 
 import './styles.css'
