+++ src/store/actionTypes.js
@@ -1,0 +1,5 @@
+export const RUN = 'RUN';
+export const STOP = 'STOP';
+export const OPEN_RANDOM = 'OPEN_RANDOM';
+export const CREATE_DISJOINT_SET = 'CREATE_DISJOINT_SET';
+export const ADD_STATS = 'ADD_STATS';
