+++ src/store/actions.js
@@ -1,6 +1,6 @@
 import {
   RUN, STOP, OPEN_RANDOM, CREATE_DISJOINT_SET, ADD_STATS
-} from './constants';
+} from './actionTypes';
 
 export const createDisjointSet = n => ({
   type: CREATE_DISJOINT_SET,
